Solved listview issues

+++ QuoteApp/QuoteApp/Globals/QuoteAppUtils.cs
@@ -68,7 +68,7 @@ namespace QuoteApp.Globals
         /// Used to get font size
         /// </summary>
         /// <returns></returns>
-        public static int PxToPt(double px)
+        public static int PxToPt(int px)
         {
             return (int)(px * 0.75);
         }
